refactor: error handling and logging; enhance AppError structure with validation details and user safety flags
- Introduced new error handling functions for common HTTP status codes (BadRequest, Unauthorized, NotFound, etc.)
- Enhanced AppError struct to include validation errors and field information
- Updated error response handling to differentiate between safe and unsafe errors for user exposure
- Replaced logrus with slog for structured logging across the application
- Removed deprecated password strength validation function
- Added JWT token validation logic with custom error handling
- Added rate limiting middleware to prevent abuse
- Improved database connection handling with error wrapping and migration checks
- Implemented a new validation package with custom validators for email and URL safety
- Refactored server initialization to use dependency injection for repositories, services, and handlers
- Updated Swagger documentation comments for clarity
- Enhances Dockerfile to specify build arguments and streamline the build context for better performance and security.

+++ api/service/season_service.go
@@ -9,10 +9,10 @@ import (
 
 type SeasonService interface {
 	CreateSeason(ctx context.Context, season *model.Season) error
-	GetSeasonByID(ctx context.Context, id uint8) (*model.Season, error)
+	GetSeasonByID(ctx context.Context, id uint) (*model.Season, error)
 	GetAllSeasons(ctx context.Context, page uint64) ([]*model.Season, error)
 	UpdateSeason(ctx context.Context, season *model.Season) error
-	DeleteSeason(ctx context.Context, id uint8) error
+	DeleteSeason(ctx context.Context, id uint) error
 }
 
 type seasonService struct {
@@ -29,7 +29,7 @@ func (s *seasonService) CreateSeason(ctx context.Context, season *model.Season) 
 	return s.SeasonRepository.CreateSeason(ctx, season)
 }
 
-func (s *seasonService) GetSeasonByID(ctx context.Context, id uint8) (*model.Season, error) {
+func (s *seasonService) GetSeasonByID(ctx context.Context, id uint) (*model.Season, error) {
 	return s.SeasonRepository.GetSeasonByID(ctx, id)
 }
 
@@ -41,6 +41,6 @@ func (s *seasonService) UpdateSeason(ctx context.Context, season *model.Season) 
 	return s.SeasonRepository.UpdateSeason(ctx, season)
 }
 
-func (s *seasonService) DeleteSeason(ctx context.Context, id uint8) error {
+func (s *seasonService) DeleteSeason(ctx context.Context, id uint) error {
 	return s.SeasonRepository.DeleteSeason(ctx, id)
 }
